+++ test/cookbooks/duounix_test/metadata.rb
@@ -1,0 +1,10 @@
+name 'duounix_test'
+maintainer 'The Authors'
+maintainer_email 'you@example.com'
+license 'All Rights Reserved'
+description 'Installs/Configures duounix_test'
+long_description 'Installs/Configures duounix_test'
+version '0.1.0'
+chef_version '>= 12.14' if respond_to?(:chef_version)
+
+depends 'duounix'